api/requirements.txt
Django==5.2.8
djangorestframework==3.15.2
django-cors-headers==4.6.0
gunicorn==23.0.0
whitenoise==6.11.0
